refactor(canon): move shard sizing into its own module
The source-file-length check rejects growth in files already over 350
lines, and the shard count helper pushed cli.rs and its test file past
it. Keep the helper and its tests together in a small sibling module
instead.
