+++ crates/vize_canon/src/batch/executor/cli/tests.rs
@@ -347,41 +347,3 @@ fn parses_materialized_tsconfig_diagnostics_as_project_level_errors() {
 
     let _ = fs::remove_dir_all(&case_dir);
 }
-
-/// With the checker count pinned to one (#3905), a process occupies one core,
-/// so the shards have to cover the machine themselves.
-#[test]
-fn shards_cover_the_machine_at_one_checker_per_process() {
-    use super::shard_count;
-
-    // 32-vCPU runner, 1000-file corpus: the old fixed divisor of 4 (plus a cap
-    // of 4) left four single-checker processes on 32 cores.
-    assert_eq!(shard_count(32, 1, 1000), 8);
-    assert_eq!(shard_count(8, 1, 1000), 8);
-    assert_eq!(shard_count(4, 1, 1000), 4);
-    assert_eq!(shard_count(1, 1, 1000), 1);
-}
-
-#[test]
-fn shard_count_accounts_for_wider_checker_pools() {
-    use super::shard_count;
-
-    // `VIZE_CHECKERS` opts back into width: four checkers per process means a
-    // quarter as many processes for the same cores.
-    assert_eq!(shard_count(32, 4, 1000), 8);
-    assert_eq!(shard_count(8, 4, 1000), 2);
-    assert_eq!(shard_count(8, 8, 1000), 1);
-    // A degenerate zero must not divide by zero.
-    assert_eq!(shard_count(8, 0, 1000), 8);
-}
-
-/// Small projects stay unsharded: an extra program's parse/bind cost is not
-/// amortized until there are enough Vue files to spread.
-#[test]
-fn small_projects_stay_unsharded() {
-    use super::shard_count;
-
-    assert_eq!(shard_count(32, 1, 0), 1);
-    assert_eq!(shard_count(32, 1, 63), 1);
-    assert_eq!(shard_count(32, 1, 128), 2);
-}
